requirements.txt
pytest
requests
allure-pytest
jsonschema
pytest-xdist